requirements.txt
click==8.1.8
Flask==2.2.5
gunicorn==20.0.4
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
Werkzeug==3.1.3
zipp==3.23.0
